src: add general::immune_level.
This allows a user to customize the prerequisite for kick immunity on their
server.

The available options are immune, admin and ircop.  If immune is set, which
is the default, the behaviour is unchanged.  If admin or ircop are set, then
the requirement of having a kick immunity mode is removed.  This may be useful
for IRCds which do not provide kick immunity modes (like inspircd).